+++ lib/widget/stateless/give_message.dart
@@ -1,0 +1,14 @@
+
+import 'package:flutter/material.dart';
+
+void giveMessage(BuildContext context, String msg) {
+  var alertDialog = AlertDialog(
+    title: Text("Message"),
+    content: Text("$msg"),
+  );
+  showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return alertDialog;
+      });
+}
